fix(case-sensitive): rename NavBar and Footer to lowercase and update tsconfig
